pyproject.toml
[build-system]
requires      = ["hatchling"]
build-backend = "hatchling.build"

[project]
name             = "kanademinder"
version          = "0.1.0"
requires-python  = ">= 3.11"
dependencies     = ["rich>=13.0", "pywebview>=5.0"]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-cov>=5.0", "ruff>=0.4", "pyinstaller>=6.0", "pillow>=10.0"]

[project.scripts]
kanademinder = "kanademinder.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/kanademinder"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
